DPUC Card lists are now a separate component
